requirements.txt
Django==1.4.1
MySQL-python==1.2.3
South==0.8.2
argparse==1.2.1
distribute==0.6.49
django-autoslug-field==0.2.3
django-debug-toolbar==0.9.4
django-google-charts==0.1.1
django-qsstats-magic==0.7.2
python-dateutil==2.1
python-nvd3==0.11.0
six==1.4.1
wsgiref==0.1.2
